测试完毕代理服务器导入功能，将extjs简化到最简化版本，二级代理初步完成，还需检查时支持post的代理还是支持get的代理，还是支持telnet的代理

+++ dahuangit-base/src/main/java/com/dahuangit/base/dao/BaseDao.java
@@ -219,8 +219,8 @@ public class BaseDao<T extends BaseModel, PK extends Serializable> {
 	 * @param hql
 	 * @return
 	 */
-	public Integer findRecordsCount(String hql) {
-		return (Integer) createQuery(hql).uniqueResult();
+	public Long findRecordsCount(String hql) {
+		return (Long) createQuery(hql).uniqueResult();
 	}
 
 	/**
